+++ traffic_ops/traffic_ops_golang/deliveryservice/eligible_test.go
@@ -1,65 +1,31 @@
-package servers
+package deliveryservice
 
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
 
 import (
 	"testing"
 
 	"github.com/apache/trafficcontrol/lib/go-tc"
 	"github.com/apache/trafficcontrol/lib/go-util"
-	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/auth"
-	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/dbhelpers"
 	"github.com/jmoiron/sqlx"
 
 	"gopkg.in/DATA-DOG/go-sqlmock.v1"
 )
 
-func TestValidateDSSAssignments(t *testing.T) {
-	expected := `server and delivery service CDNs do not match`
-	cdnID := 1
-	ds := DSInfo{
-		ID:    0,
-		CDNID: &cdnID,
-	}
-	var servers []dbhelpers.ServerHostNameCDNIDAndType
-	server := dbhelpers.ServerHostNameCDNIDAndType{
-		HostName: "serverHost",
-		CDNID:    0,
-		Type:     "",
-	}
-	servers = append(servers, server)
-	userErr := ValidateDSSAssignments(ds, servers)
-	if userErr == nil {
-		t.Fatalf("Expected user error with mismatching ds and server CDN IDs, got no error instead")
-	}
-	if userErr.Error() != expected {
-		t.Errorf("Expected error details %v, got %v", expected, userErr.Error())
-	}
-	servers[0].CDNID = 1
-	userErr = ValidateDSSAssignments(ds, servers)
-	if userErr != nil {
-		t.Fatalf("Expected no user error, got %v", userErr.Error())
-	}
-}
-
-func TestReadServers(t *testing.T) {
+func TestGetEligibleServers(t *testing.T) {
 	mockDB, mock, err := sqlmock.New()
 	if err != nil {
 		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
@@ -103,7 +69,9 @@ func TestReadServers(t *testing.T) {
 		"tcp_port",
 		"server_type",
 		"server_type_id",
-		"upd_pending"}
+		"upd_pending",
+		"server_capabilities",
+		"deliveryservice_capabilities"}
 
 	rows := sqlmock.NewRows(cols)
 
@@ -143,6 +111,8 @@ func TestReadServers(t *testing.T) {
 			s.Type,
 			s.TypeID,
 			s.UpdPending,
+			[]byte(`{""}`),
+			[]byte(`{""}`),
 		)
 	}
 
@@ -150,7 +120,7 @@ func TestReadServers(t *testing.T) {
 	mock.ExpectQuery("SELECT").WillReturnRows(rows)
 	mock.ExpectCommit()
 
-	actualSrvs, err := read(db.MustBegin(), 1, &auth.CurrentUser{PrivLevel: 30}, false)
+	actualSrvs, err := getEligibleServers(db.MustBegin().Tx, 1)
 	if err != nil {
 		t.Fatalf("an error '%s' occurred during read", err)
 	}